validação para salvar somente quando arquivo já não estiver no servidor

+++ backend/reurb/src/main/java/br/com/cronos/appreurb/controller/DadosPessoaisController.java
@@ -24,9 +24,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import br.com.cronos.appreurb.model.ArquivoDocumentoPessoal;
 import br.com.cronos.appreurb.model.DadosConjuge;
 import br.com.cronos.appreurb.model.DadosPessoais;
 import br.com.cronos.appreurb.model.IntegranteImovel;
+import br.com.cronos.appreurb.repository.ArquivoDocumentoPessoalRepository;
 import br.com.cronos.appreurb.repository.DadosConjugeRepository;
 import br.com.cronos.appreurb.repository.DadosPessoaisRepository;
 import br.com.cronos.appreurb.repository.IntegranteImovelRepository;
@@ -48,6 +50,9 @@ public class DadosPessoaisController
 	
 	@Autowired
 	IntegranteImovelRepository integranteImovelRepository;
+	
+	@Autowired
+	ArquivoDocumentoPessoalRepository arquivoDocumentoPessoalRepository;
 	
 	@RequestMapping(value="/dados_pessoais", method= RequestMethod.GET)
 	public ResponseEntity<List<DadosPessoais>> listarDadosPessoais() 
@@ -165,18 +170,32 @@ public class DadosPessoaisController
 	@PostMapping("dados_pessoais/arquivoPessoal/{id}")
     public ResponseEntity<String> salvarArquivo(@PathVariable("id") Long id, @RequestParam("file") MultipartFile file) 
 	{
-		System.out.println("Recebendo o arquivo: " + file.getOriginalFilename());
+		//System.out.println("Recebendo o arquivo: " + file.getOriginalFilename());
 		
 //		String caminho = pathArquivos + UUID.randomUUID() + "-" + file.getOriginalFilename();
-		String caminho = pathArquivos + id + "-" + file.getOriginalFilename();
+		String novoNomeArquivo = id + " --- " + file.getOriginalFilename();
+		String caminho = pathArquivos + novoNomeArquivo;
 
-		System.out.println("Novo nome do arquivo: " + caminho);
+		System.out.println("Novo nome do arquivo: " + novoNomeArquivo);
+		System.out.println("Caminho pro arquivo: " + caminho);
 		
 		Path path = FileSystems.getDefault().getPath(caminho);
 
-        try {
-            Files.copy(file.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
-            
+        try 
+        {
+        	System.out.println(Files.exists(path));
+        	if(!Files.exists(path))
+        	{
+        		Files.copy(file.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
+                
+                ArquivoDocumentoPessoal arquivoDocumentoPessoal = new ArquivoDocumentoPessoal();
+                arquivoDocumentoPessoal.setNomeArquivo(novoNomeArquivo);
+                DadosPessoais dadosPessoais = new DadosPessoais();
+                dadosPessoais.setId(id);
+                arquivoDocumentoPessoal.setDadosPessoais(dadosPessoais);
+                arquivoDocumentoPessoalRepository.save(arquivoDocumentoPessoal);
+        	}
+        	
             return new ResponseEntity<>("{ \"mensagem\": \"Arquivo carregado com sucesso!\"}", HttpStatus.OK);
         } catch (Exception e) {
         	e.printStackTrace();
